% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno.orgHs}
\alias{anno.orgHs}
\title{annotation for translating between gene identifiers}
\format{An object of class \code{data.frame} with 63242 rows and 6 columns.}
\usage{
anno.orgHs
}
\description{
Table for translating between Ensembl, NCBI Entrez and HGNC symbol
human gene identifiers. Gene biotype, exonic gene length and proportion
GC-bases are also included. For more exhaustive annotation including gene
names, GO terms and so on, please refer to R Bioconductor packages
\code{AnnotationDbi} and \code{org.Hs.eg.db}.
}
\details{
Gene identifiers are from \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
package. Gene biotype is from Ensembl biomaRt
(\url{http://www.ensembl.org/biomart}). Gene length
(exons) and proportion GC-bases were calculated based on GENCODE
v25/GRCh38.p7.genome data (\url{http://www.gencodegenes.org/}).
}
\note{
Exon gene length may be longer than any specific transcript as
all annotated exons are included.
}
\examples{
head(anno.orgHs)
dim(anno.orgHs)
colSums(apply(anno.orgHs, 2, is.na))
colSums(!apply(anno.orgHs[,1:4], 2, duplicated))
# longest gene
anno.orgHs[which.max(anno.orgHs$length),]
}
\references{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.3.0.

Durinck S, Spellman PT, Birney E, Huber W. Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt. Nat Protocols. 2009;4:1184-91.

Durinck S, Moreau Y, Kasprzyk A, Davis S, Moor BD, Brazma A, et al. BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis. Bioinformatics. 2005;21:3439-40.

Harrow J, Frankish A, Gonzalez JM, Tapanari E, Diekhans M, Kokocinski F, et al. GENCODE: The reference human genome annotation for The ENCODE Project. Genome Res. 2012;22:1760-74.
}
\seealso{
\code{\link{fromTo}}, \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
}
\keyword{datasets}
