% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMScaller.R
\name{CMScaller}
\alias{CMScaller}
\title{CMS classification}
\usage{
CMScaller(
  emat,
  templates = CMScaller::templates.CMS,
  rowNames = "entrez",
  RNAseq = FALSE,
  nPerm = 1000,
  seed = NULL,
  FDR = 0.05,
  doPlot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{emat}{a numeric expression matrix with sample columns, gene rows and
Entrez rownames. Microarray data should be normalized. For RNA-seq data,
counts or RSEM values could be used directly by setting \code{RNAseq=TRUE}.}

\item{templates}{a data frame with two columns; \emph{class} (coerced to
factor) and \emph{probe} (coerced to character).}

\item{rowNames}{a character, either "entrez" (NCBI Entrez),
"symbol" (HGNC symbol) or "ensg" (Ensembl). If set to other than "ensg",
\code{\link{replaceGeneId}} is used to translate \code{rownames(emat)}.}

\item{RNAseq}{a logical, set to TRUE if emat is untransformed, non-normalized
sequencing counts or RSEM values.}

\item{nPerm}{an integer, number of permutations for \eqn{p}-value
estimation.}

\item{seed}{an integer, for \eqn{p}-value reproducibility.}

\item{FDR}{a false discovery rate, sets prediction confidence threshold.}

\item{doPlot}{a logical, whether to produce prediction \code{\link{subHeatmap}}.}

\item{verbose}{a logical, whether console messages are to be displayed.}
}
\value{
a data frame with class predictions, template distances,
\eqn{p}-values and false discovery rate adjusted \eqn{p}-values
(\code{\link[stats]{p.adjust}}). Rownames equal emat
colnames.
}
\description{
Cancer-cell intrinsic CMS classification based on pre-defined
subtype templates.
}
\details{
\code{CMScaller} provides classification based on pre-defined
cancer-cell intrinsic CMS templates. If \code{RNA-seq=TRUE}, a pseudocount
of 0.25 is added, matrix log2-transformed and quantile normalized
(\code{\link[limma]{normalizeQuantiles}}) prior to scaling/centering and
prediction. The core algorithm is the Nearest Template Prediction (NTP)
algorithm as proposed by Yujin Hoshida (2010). See \code{\link{ntp}} for
further details.
}
\note{
genes with missing values are discarded.
}
\examples{
res <- CMScaller(crcTCGAsubset, RNAseq=TRUE)
head(res)
hist(res$p.value)
}
\references{
Hoshida, Y. (2010). Nearest Template Prediction: A
Single-Sample-Based Flexible Class Prediction with Confidence Assessment.
PLoS ONE 5, e15543.

Guinney J, Dienstmann R, Wang X, de Reynies A, Schlicker A,
Soneson C, et al. The consensus molecular subtypes of colorectal cancer.
Nat Med. 2015;21:1350-6.
}
\seealso{
\code{\link{ntp}}, \code{\link{templates.CMS}},
\code{\link{corCosine}}
}
