% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crcTCGAsubset}
\alias{crcTCGAsubset}
\title{colorectal TCGA gene expression data with subtype annotation}
\format{
An object of class \code{ExpressionSet} with 16863 rows and 92 columns.
}
\usage{
crcTCGAsubset
}
\description{
The Cancer Genome Atlas level 3 RSEM genes data was retrieved
from Broad GDAC Firehose, doi:10.7908/C11G0KM9
(\url{http://gdac.broadinstitute.org/}). Samples were annotated with
Consensus Molecular Subtypes (Guinney 2015), Sadanandam subtypes
(Isella 2015) and ABSOLUTE estimated tumor purity (Aran 2015). Features were
annotated using \code{\link[CMScaller]{fromTo}}. Rownames are Entrez ids.
Included are 92 HiSeq test set samples with genes with maximum count
exceeding 25.
}
\examples{
library(Biobase)
dim(crcTCGAsubset)
}
\references{
Aran D, Sirota M, Butte AJ. Systematic pan-cancer analysis of tumour purity. Nat Commun. 2015;6:8971.

Guinney J, Dienstmann R, Wang X, de Reynies A, Schlicker A, Soneson C, et al. The consensus molecular subtypes of colorectal cancer. Nat Med. 2015;21:1350-6.

Isella C, Terrasi A, Bellomo SE, Petti C, Galatola G, Muratore A, et al. Stromal contribution to the colorectal cancer transcriptome. Nat Genet [Internet]. 2015 [cited 2015 Mar 2];advance online publication. Available from: \url{http://www.nature.com/ng/journal/vaop/ncurrent/full/ng.3224.html}

TCGA. Comprehensive molecular characterization of human colon and rectal cancer. Nature. 2012;487:330-7.
}
\keyword{datasets}
