% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromTo.R
\name{fromTo}
\alias{fromTo}
\title{translation between human gene identifiers}
\usage{
fromTo(
  key = NULL,
  id.in = NULL,
  id.out = "symbol",
  verbose = getOption("verbose"),
  rough = FALSE,
  all = FALSE
)
}
\arguments{
\item{key}{a character vector, gene identifiers to be translated.}

\item{id.in}{a character, \code{colnames(anno.orgHs)} are valid options.}

\item{id.out}{a character, \code{colnames(anno.orgHs)} are valid options.}

\item{verbose}{logical, whether console information are to returned.}

\item{rough}{logical, whether to search for more than one match. If FALSE,
provides less informative warnings.}

\item{all}{logical, whether to include all possible matches.
\code{length(key) > 10000}, setting \code{rough=TRUE} increases
speed by >100x, but also suppresses warnings.}
}
\value{
A character vector of same length as \emph{key} with gene
identifiers as specified.
}
\description{
Translation between human gene identifiers. For input keys with
multiple valid matches, \bold{only first match is returned} (warning is
issued). If no \code{id.in} is provided, function will make an informed
guess.
}
\details{
Gene identifiers are from the \code{org.Hs.eg.db} package while
\emph{biotype} were retrieved from \code{biomaRt} using
\code{useMart(biomart = "ensembl", dataset = "hsapiens_gene_ensembl"}.
}
\note{
Warnings are issued for keys with none or more than one hit. For keys
with more than one possible match, only first match is returned (thus
\emph{rough}).
}
\examples{
 library(Biobase)
 fromTo(rownames(crcTCGAsubset)[1:50], "entrez", "symbol")
 # colnames indicate valid keys
 # example: AKT3 symbol has one valid Entrez mapping but two different ENSG
 anno.orgHs[anno.orgHs$symbol == "AKT3",]
 fromTo("AKT3", "symbol", "ensg") # expect one
 fromTo("AKT3", "symbol", "entrez") # expect one
 fromTo("AKT3", "symbol", "ensg", rough=FALSE) # expect one
 fromTo("AKT3", "symbol", "ensg", all = TRUE) # expect two
 # number of ids with multiple mappings
 sum(table(anno.orgHs$symbol) > 1)
 sum(table(anno.orgHs$entrez) > 1)
 sum(table(anno.orgHs$ensg) > 1)
}
\references{
Marc Carlson (). org.Hs.eg.db: Genome wide annotation for Human.
R package version 3.2.3.

Steffen Durinck et al. (2009) Mapping identifiers for the
integration of genomic datasets with the R/Bioconductor package. Nature
Protocols 4, 1184-1191.
}
\seealso{
\code{\link{anno.orgHs}}, \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
}
