% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneSets.CMS}
\alias{geneSets.CMS}
\title{gene sets relevant to Consensus Molecular Subtypes}
\format{
An object of class \code{list} of length 14.
}
\usage{
geneSets.CMS
}
\description{
Geneset is a named list of Entrez ids and is a subset of \code{\link{geneSets.CRC}}.
\itemize{
\item{MSI (Watanabe) from \href{http://software.broadinstitute.org/gsea/msigdb}{MutSigDB} C2 (v7.0)}
\item{DNA repair from MutSigDB Hallmark (v7.0)}
\item{HNF4A (Lucas) from MutSigDB C2 (v7.0)}
\item{MSS (Watanabe) from MutSigDB C2 (v7.0)}
\item{MYC from MutSigDB Hallmark (v7.0)}
\item{WNT signature based on Vermeulen Supplementary Table S1}
\item{cell cycle (E2F targets) from MutSigDB Hallmark (v7.0)}
\item{differentiation (gastro-intestinal markers) from
\href{http://www.proteinatlas.org/humanproteome/gastrointestinal+tract}{Human Protein Atlas}.}
\item{glycolysis from MutSigDB Hallmark (v7.0)}
\item{fatty acids from \href{http://www.reactome.org/}{reactome} (accessed 20190822)}
\item{CDX2 (Liu) from MutSigDB C2 (v7.0)}
\item{LGR5 stem cells Merlos-Suarez Supplementary Table 5}
\item{TGF-beta signature based on Fessler GEO \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE79461}{GSE79461}}
\item{EMT (epithelial-mesenchymal transition)from MutSigDB Hallmark (v7.0)}
}
}
\examples{
names(geneSets.CMS)
}
\references{
Croft D, Mundo AF, Haw R, Milacic M, Weiser J, Wu G, et al. The Reactome pathway knowledgebase. Nucl Acids Res. 2014;42:D472-7.

Fessler E, Drost J, Hooff SR van, Linnekamp JF, Wang X, Jansen M, et al. TGF-beta signaling directs serrated adenomas to the mesenchymal colorectal cancer subtype. EMBO Molecular Medicine. 2016;8:745-60.

Liu T, Zhang X, So C-K, Wang S, Wang P, Yan L, et al. Regulation of Cdx2 expression by promoter methylation, and effects of Cdx2 transfection on morphology and gene expression of human esophageal epithelial cells. Carcinogenesis. 2007;28:488-96.

Lucas B, Grigo K, Erdmann S, Lausen J, Klein-Hitpass L, Ryffel GU. HNF4alpha reduces proliferation of kidney cells and affects genes deregulated in renal cell carcinoma. Oncogene. 2005;24:6418-31.

Guinney J, Dienstmann R, Wang X, de Reynies A, Schlicker A, Soneson C, et al. The consensus molecular subtypes of colorectal cancer. Nat Med. 2015;21:1350-6.

Liberzon A, Subramanian A, Pinchback R, Thorvaldsdottir H, Tamayo P, Mesirov JP. Molecular signatures database (MSigDB) 3.0. Bioinformatics. 2011;27:1739-40.

Merlos-Suarez A, Barriga FM, Jung P, Iglesias M, Cespedes MV, Rossell D, et al. The Intestinal Stem Cell Signature Identifies Colorectal Cancer Stem Cells and Predicts Disease Relapse. Cell Stem Cell. 2011;8:511-24.

Uhlen M, Fagerberg L, Hallstrom BM, Lindskog C, Oksvold P, Mardinoglu A, et al. Tissue-based map of the human proteome. Science. 2015;347:1260419.

Vermeulen L, De Sousa E Melo F, van der Heijden M, Cameron K, de Jong JH, Borovski T, et al. Wnt activity defines colon cancer stem cells and is regulated by the microenvironment. Nature Cell Biology. 2010;12:468-76.

Watanabe T, Kobunai T, Toda E, Yamamoto Y, Kanazawa T, Kazama Y, et al. Distal colorectal cancers with microsatellite instability (MSI) display distinct gene expression profiles that are different from proximal MSI cancers. Cancer Res. 2006;66:9804-8.
}
\keyword{datasets}
