% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneSets.CRC}
\alias{geneSets.CRC}
\title{gene sets relevant to Consensus Molecular Subtypes}
\format{
An object of class \code{list} of length 69.
}
\usage{
geneSets.CRC
}
\description{
Geneset is a named list of Entrez idds. Watanabe CRC MSS/MSI,
Liu CDX2, Lucas HNF4A, and Servitja HNF1A were retrieved from MutSigDB C2
(v7.0). Gastro-Intestinal enriched genes are from the \href{http://www.proteinatlas.org/humanproteome/gastrointestinal+tract}{Human Protein Atlas}
Extracellular Matrix mCRC is from Naba Additional Data 2.
Crypt signatures are based on Merlos-Suarez Supplementary Table 5.
MYB signature is based on Thorner Supplementary Table 2.
Stromal estimate is based on Yoshihara Supplementary Data 1.
CRC stem cell signatures are based on de Sousa E Melo Supplementary Table 1A.
TGF-beta signatures are based on reanalysis of Fessler GSE79461.
CTNNB1/Beta-catenin signature is based on Watanabe Supplementary Table 1.
WNT signatures are based on Vermeulen Supplementary Table S1.
Retionic acid signatures are based on Duffy Supplementary Table 2.
Remaining are either from MutSigDB Hallmarks (v7.0) or Reactome (Aug-2019).
}
\examples{
names(geneSets.CRC)
}
\references{
Croft D, Mundo AF, Haw R, Milacic M, Weiser J, Wu G, et al. The Reactome pathway knowledgebase. Nucl Acids Res. 2014;42:D472-7.

Duffy DJ, Krstic A, Halasz M, Schwarzl T, Konietzny A, Iljin K, et al. Retinoic acid and TGF-beta signalling cooperate to overcome MYCN-induced retinoid resistance. Genome Medicine. 2017;9:15.

Fessler E, Drost J, Hooff SR van, Linnekamp JF, Wang X, Jansen M, et al. TGF-beta signaling directs serrated adenomas to the mesenchymal colorectal cancer subtype. EMBO Molecular Medicine. 2016;8:745-60.

Guinney J, Dienstmann R, Wang X, de Reynies A, Schlicker A, Soneson C, et al. The consensus molecular subtypes of colorectal cancer. Nat Med. 2015;21:1350-6.

Liberzon A, Subramanian A, Pinchback R, Thorvaldsdottir H, Tamayo P, Mesirov JP. Molecular signatures database (MSigDB) 3.0. Bioinformatics. 2011;27:1739-40.

Merlos-Suarez A, Barriga FM, Jung P, Iglesias M, Cespedes MV, Rossell D, et al. The Intestinal Stem Cell Signature Identifies Colorectal Cancer Stem Cells and Predicts Disease Relapse. Cell Stem Cell. 2011;8:511-24.

Naba A, Clauser KR, Whittaker CA, Carr SA, Tanabe KK, Hynes RO. Extracellular matrix signatures of human primary metastatic colon cancers and their metastases to liver. BMC Cancer. 2014;14:518.

de Sousa E Melo F, Colak S, Buikhuisen J, Koster J, Cameron K, de Jong JH, et al. Methylation of Cancer-Stem-Cell-Associated Wnt Target Genes Predicts Poor Prognosis in Colorectal Cancer Patients. Cell Stem Cell. 2011;9:476-85.

Thorner AR, Parker JS, Hoadley KA, Perou CM. Potential Tumor Suppressor Role for the c-Myb Oncogene in Luminal Breast Cancer. PLoS One [Internet]. 2010 [cited 2017 Jan 19];5. Available from: http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2951337/

Uhlen M, Fagerberg L, Hallstrom BM, Lindskog C, Oksvold P, Mardinoglu A, et al. Tissue-based map of the human proteome. Science. 2015;347:1260419.

Vermeulen L, De Sousa E Melo F, van der Heijden M, Cameron K, de Jong JH, Borovski T, et al. Wnt activity defines colon cancer stem cells and is regulated by the microenvironment. Nature Cell Biology. 2010;12:468-76.

Watanabe K, Biesinger J, Salmans ML, Roberts BS, Arthur WT, Cleary M, et al. Integrative ChIP-seq/Microarray Analysis Identifies a CTNNB1 Target Signature Enriched in Intestinal Stem Cells and Colon Cancer. PLoS One [Internet]. 2014 [cited 2017 Jan 19];9. Available from: http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3961325/

Yoshihara K, Shahmoradgoli M, Martinez E, Vegesna R, Kim H, Torres-Garcia W, et al. Inferring tumour purity and stromal and immune cell admixture from expression data. Nat Commun. 2013;4:2612.
}
\keyword{datasets}
