% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneSets.reactome}
\alias{geneSets.reactome}
\title{gene sets for exploratory gene set analysis}
\format{
An object of class \code{list} of length 2247.
}
\usage{
geneSets.reactome
}
\description{
gene sets for exploratory gene set analysis
}
\details{
Gene sets from Reactome (Aug-2019) with Entrez gene identifiers.
}
\examples{
head(names(geneSets.reactome))
}
\references{
Croft D, Mundo AF, Haw R, Milacic M, Weiser J, Wu G, et al. The Reactome pathway knowledgebase. Nucl Acids Res. 2014;42:D472-7.
}
\keyword{datasets}
