% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmCMScaller.R
\name{lmCMScaller}
\alias{lmCMScaller}
\title{liver metastases CMS classification}
\usage{
lmCMScaller(emat, rowNames = "entrez", RNAseq = FALSE, posterior = 0.5)
}
\arguments{
\item{emat}{a numeric expression matrix with sample columns, gene rows and
Entrez rownames. Microarray data should be normalized. For RNA-seq data,
counts or RSEM values could be used directly by setting \code{RNAseq=TRUE}.}

\item{rowNames}{a character, either "entrez" (NCBI Entrez),
"symbol" (HGNC symbol) or "ensg" (Ensembl). If set to other than "ensg",
\code{\link{replaceGeneId}} is used to translate \code{rownames(emat)}.}

\item{RNAseq}{a logical, set to TRUE if emat is untransformed, non-normalized
sequencing counts or RSEM values.}

\item{posterior}{numeric, sets minimum prediction confidence threshold.}
}
\value{
a data frame with columns of class-wise posterior probabilities,
predictions and maximum posterior probilities. Rownames equal \code{emat}
colnames.
}
\description{
liver metastases CMS classification
}
\details{
\code{lmCMScaller} provides CMS classification of colorectal cancer
liver metastasis based on a pre-defined \code{\link[randomForest]{randomForest}}
model. If \code{RNA-seq=TRUE}, a pseudocount of 0.25 is added, matrix
log2-transformed and quantile normalized (\code{\link[limma]{normalizeQuantiles}})
prior to scaling/centering and prediction.
}
\examples{
library(randomForest)
data("mcrcOSLOsubset")
emat <- mcrcOSLOsubset[,!duplicated(mcrcOSLOsubset$`Patient ID`)]
res1 <- lmCMScaller(emat)
sum(is.na(res1))
head(res1)
table(res1$prediction, emat$cms_pam)
## see that it handles (some) missing values
res2 <- lmCMScaller(emat[-c(1:100),])
sum(is.na(res2))
table(res1$prediction, res2$prediction)
}
\references{
Guinney J, Dienstmann R, Wang X, de Reyniès A, Schlicker A, Soneson C, et al. The consensus molecular subtypes of colorectal cancer. Nat Med. 2015;21:1350–6. Available from: \url{https://www.nature.com/articles/nm.3967}

Liaw A, Wiener M. Classification and Regression by randomForest. R News. 2002;2:18–22. Available from: \url{https://cran.r-project.org/doc/Rnews/Rnews_2002-3.pdf}
}
