% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcrcOSLOsubset}
\alias{mcrcOSLOsubset}
\title{colorectal cancer liver metastasis gene expression data}
\format{
An object of class \code{ExpressionSet} with 1373 rows and 295 columns.
}
\usage{
mcrcOSLOsubset
}
\description{
Fresh-frozen resected colorectal cancer liver metastaseses were
profiled using Affymetrix HTA 2.0 microarrays. Data was preprocessed using
\code{\link[affy]{justRMA}} (v1.66.0) with \href{http://brainarray.mbni.med.umich.edu/Brainarray/Database/CustomCDF/CDF_download.asp}{BRAINARRAY} (v24) ENTREZ CDF (Sandberg & Lasson). Raw data
will be available from Gene Expression Omnibus.
}
\details{
Only protein-coding genes used for CMS prediction are included.
Features were annotated using \code{\link[CMScaller]{fromTo}}. \code{rownames}
are ENTREZ identifiers.
}
\examples{
library(Biobase)
dim(mcrcOSLOsubset)
}
\references{
Sandberg R, Larsson O. Improved precision and accuracy for microarrays using updated probe set definitions. BMC Bioinform. 2007;8:1. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1805763/}
}
\keyword{datasets}
