% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntp.R
\name{ntp}
\alias{ntp}
\title{nearest template prediction}
\usage{
ntp(
  emat,
  templates,
  nPerm = 1000,
  distance = "cosine",
  nCores = 1,
  seed = NULL,
  verbose = getOption("verbose"),
  doPlot = FALSE
)
}
\arguments{
\item{emat}{a numeric matrix with row features and sample columns.
\code{rownames(emat)} are matched against \code{templates$probe}.}

\item{templates}{a data frame with two columns; \emph{class} (coerced to
factor) and \emph{probe} (coerced to character).}

\item{nPerm}{an integer, number of permutations for \eqn{p}-value
estimation.}

\item{distance}{a character, either c("cosine", "pearson", "spearman" or
"kendall").}

\item{nCores}{an integer specifying number of threads for
parallelization.}

\item{seed}{an integer, for \eqn{p}-value reproducibility. Setting seed
enforces serial processing.}

\item{verbose}{logical, whether console messages are to be displayed.}

\item{doPlot}{logical, whether to produce prediction \code{\link{subHeatmap}}.}
}
\value{
a data frame with class predictions, template distances,
\eqn{p}-values and false discovery rate adjusted \eqn{p}-values
(\code{\link[stats]{p.adjust}(method = "fdr")}). Rownames equal emat
colnames.
}
\description{
Nearest Template Prediction (NTP) based on predefined class
templates.
}
\details{
\code{ntp} implements the Nearest Template Prediction (NTP)
algorithm, largely as proposed by Yujin Hoshida (2010) (see below). For each
sample, distances to templates are calculated and class assigned based on
smallest distance. Distances are transformed from the sample-templates
correlations as follows:
\deqn{d.class = \sqrt(1/2 * (1-(cor(sample,templates))}
Template values are 1 for class features and 0 for non-class features (-1 if
there are only two classes). Prediction confidence is estimated based on
the distance of the null-distribution, estimated from permutation tests.
Thus the lowest possible estimate of the \eqn{p}-value is \eqn{1/nPerm}.
\itemize{
 \item{\code{emat}}{ should be a row-wise \emph{centered and scaled} matrix.
  For large, balanced datasets, this may be achieved by applying
  \code{\link{ematAdjust}} function.}
 \item{\code{templates}}{ is a data.frame defining class templates. A class
  template is a set of marker genes with higher expected expression in
  samples belonging to class compared to non-class samples. \code{templates}
  must contain at least two columns named \emph{probe} and \emph{class}.}
 \item{compared to Hoshida (2010), resulting \eqn{p}-value estimates are more
 conservative (by a factor equaling the number of classes) and
 the distances are a monotonic transformation of \eqn{1-cor} (see
 Details section above).}
 \item{Hoshida (2010) does not explicitly state whether input should be
 log2-transformed or not and examples includes both. Based on experience
 this choice affects results only at the margins, but for high-quality
 datasets, normalized, untransformed inputs may yield a small increase in
 accuracy.}
 }
For further details on the NTP algorithm, please refer to package vignette
and \href{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0015543}{Hoshida (2010)}.

Parallel processing is implemented through \pkg{parallel}
\code{\link[parallel]{mclapply}} or \pkg{snow} \code{\link[snow]{parLapply}}
for nix and Windows systems, respectively.
}
\note{
\itemize{
\item{features with missing values are discarded.}
\item{setting seed disables parallel processing to ensure \eqn{p}-value
reproducibility.}
\item{for two random uncorrelated vectors \eqn{x,y} \eqn{N\sim(0,1)}
\eqn{E[d.xy]\approx0.71} when distance is cosine.}
\item{internally, correlations instead of distances are calculated.}
\item{accepts reuse of features (marker not specific for one class only)}}
}
\examples{
emat <- ematAdjust(crcTCGAsubset, normMethod = "quantile")
res <- ntp(emat, templates.CMS, doPlot=TRUE, nPerm=100)
head(res)
}
\references{
Hoshida, Y. (2010). Nearest Template Prediction: A
Single-Sample-Based Flexible Class Prediction with Confidence Assessment.
PLoS ONE 5, e15543.
}
\seealso{
\code{\link{corCosine}}, \code{\link[stats]{cor}}
}
