% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPC.R
\name{plotPC}
\alias{plotPC}
\title{2d plot of PCA variable loadings}
\usage{
plotPC(x, dim = c(1:2), n = 10, entrez = FALSE, ...)
}
\arguments{
\item{x}{list, PCA output (see example code)}

\item{dim}{a numeric vector of length 2, specifying which principal
components to plot.}

\item{n}{numeric, number of variables to include}

\item{entrez}{logical, if TRUE, translates variables from Entrez to Gene symbosl}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}}}
}
\description{
Plots the loadings for the highest contributing variables (by euclidean distance).
}
\examples{
library(CMScaller)
p <- subPCA(crcTCGAsubset[,1:10])
plotPC(p, entrez=TRUE)
}
\seealso{
\code{\link[stats]{prcomp}}, \code{\link{subPCA}}
}
