% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceGeneId.R
\name{replaceGeneId}
\alias{replaceGeneId}
\title{replace matrix gene identifiers}
\usage{
replaceGeneId(emat, id.in = "symbol", id.out = "entrez")
}
\arguments{
\item{emat}{a numeric matrix with row features and sample columns.}

\item{id.in}{a character, gene symbol, entrez and ENSG are valid options.}

\item{id.out}{a character, gene symbol, entrez and ENSG are valid options.}
}
\value{
A matrix with rownames gene identifier as specified in id.out.
For duplicates, rownames for rows with lower standard deviation are added a
number to make unique.
\code{\link{make.unique}} are used to handle NA and duplicates in output.
}
\description{
Translate matrix rownames between gene identifiers.
}
\details{
Gene identifiers are from the
\code{\link[org.Hs.eg.db]{org.Hs.eg.db}} package.
}
\examples{
# example data
library(Biobase)
x.entrez <- Biobase::exprs(crcTCGAsubset)
# rownames are entrez
head(rownames(x.entrez))
# translate to gene symbols
x.symbol <- replaceGeneId(x.entrez, id.in="entrez", id.out="symbol")
head(rownames(x.symbol))
x.entrez2 <- replaceGeneId(x.symbol, id.in="symbol", id.out="entrez")
# translations are not cycle consistent
table(rownames(x.entrez2) == rownames(x.entrez))
# matrix values are not changed
all(x.entrez == x.entrez2)
}
\references{
Marc Carlson (). org.Hs.eg.db: Genome wide annotation for Human.
R package version 3.2.3.

Steffen Durinck et al. (2009) Mapping identifiers for the
integration of genomic datasets with the R/Bioconductor package. Nature
Protocols 4, 1184-1191.
}
\seealso{
\code{\link{anno.orgHs}}, \code{\link{fromTo}},
\code{\link[org.Hs.eg.db]{org.Hs.eg.db}}, \code{\link[stats]{sd}}
\code{\link{make.unique}}
}
