% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subBoxplot.R
\name{subBoxplot}
\alias{subBoxplot}
\title{boxplot with data points}
\usage{
subBoxplot(
  y,
  x,
  labels = NULL,
  keepN = TRUE,
  classCol = getOption("subClassCol"),
  points = FALSE,
  width = 20,
  showOutliers = FALSE,
  ...
)
}
\arguments{
\item{y}{a numeric vector.}

\item{x}{a factor, \code{length(x)==length(y)}.}

\item{labels}{a character vector, if NULL, \code{names(y)} is used.}

\item{keepN}{logical or numeric indicating which samples to include.}

\item{classCol}{a character vector specifying class colors.}

\item{points}{logical indicating whether points are to be added to boxes.}

\item{width}{numeric to adjust point stacking - increase to reduce width.}

\item{showOutliers}{logical indicating whether outliers are highlighted.}

\item{...}{additional arguments to be passed to \code{graphics::boxplot}.}
}
\value{
a boxplot
}
\description{
Boxplot with data points and labeled outliers.
}
\note{
For data points, \eqn{y}-values are grouped and averaged - not exact!
}
\examples{
isCMS4 <- rownames(crcTCGAsubset) \%in\%
    templates.CMS$probe[templates.CMS$class == 'CMS4']
y <- colSums(Biobase::exprs(crcTCGAsubset)[isCMS4,])
x <- crcTCGAsubset$CMS
subBoxplot(y, x, ylab = expression(sum(log2(RSEM))), main = 'CMS4 genes',
     keepN = which(x != "CMS1"), notch = TRUE)
}
