% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subCamera.R
\name{subCamera}
\alias{subCamera}
\title{limma/voom camera gene set analysis visualization}
\usage{
subCamera(
  emat,
  class,
  keepN = TRUE,
  batch = NULL,
  geneList = NULL,
  xKey = NULL,
  pValue = 0.01,
  topN = 15,
  allowNegCor = FALSE,
  interGeneCor = 0.01,
  doPlot = TRUE,
  doVoom = FALSE,
  normMethod = "quantile",
  pMax = 10,
  rowCluster = TRUE,
  classCol = getOption("subClassCol"),
  heatCol = NULL,
  cexText = 1,
  legendAdd = TRUE,
  ...
)
}
\arguments{
\item{emat}{numeric matrix with row features and sample columns.}

\item{class}{a factor vector specifying sample classes
\code{length(class)==ncol(emat)}.}

\item{keepN}{a logical or numeric vector specifying which samples to keep
(defaults to all).}

\item{batch}{a factor vector specifying additional level in design matrix.}

\item{geneList}{a named list. where each item is a gene set consisting of a
character vector of genes belonging  to that set.}

\item{xKey}{a character vector matching emat with geneList identifiers.
\code{length(xKey)==nrow(emat)}. Leave blank if rownames(emat) and geneList
uses same identifiers.}

\item{pValue}{a p-value (number) indicating minimum gene set significance
included in plot.}

\item{topN}{integer, number of gene sets to include in the plot.
(for K>2, lowest p-value for each class is included).}

\item{allowNegCor}{logical, passed to limma::\code{\link[limma]{camera}} as
allow.neg.cor parameter.}

\item{interGeneCor}{a number, passed to limma::\code{\link[limma]{camera}}
as inter.gene.cor parameter.}

\item{doPlot}{a logical, indicating whether to return plot (TRUE) or
significance values (FALSE) of results.}

\item{doVoom}{a logical, indicating whether emat is \emph{untransformed}
sequencing count data and \code{\link[limma]{voom}} should be used for
modeling.}

\item{normMethod}{a character, only used if doVoom=TRUE and passed to
\code{\link[edgeR]{calcNormFactors}} if element in c("TMM","RLE",
"upperquartile","none") or \code{\link[limma]{voom}} ("scale", "quantile",
"cyclicloess").}

\item{pMax}{a numeric, \eqn{log10(-pValue)} for color scale.}

\item{rowCluster}{logical, indicating whether heatmap rows should be
clustered.}

\item{classCol}{a character vector specifying class colors.}

\item{heatCol}{a character vector specifying heatmap colors.}

\item{cexText}{numeric, text scaling factor.}

\item{legendAdd}{a logical, whether to add legend to barplot.}

\item{...}{additional arguments passed to \code{\link[graphics]{image}} or
in case of \code{levels(class)==2} \code{\link[graphics]{barplot}}.}
}
\value{
a barplot/heatmap and \code{\link[limma]{camera}} output (list,
invisible). For two-classes comparison, gene sets are ranked by significance
and bars are colored according to relative up-regulation. In heatmap with
default \code{heatCol}, red and blue indicates relative up- and
down-regulation respectively. Color intensity reflects significance.
Nominal `camera` \eqn{p}-values are used as input for visualization.
}
\description{
Wrapper function to perform \code{\link[limma]{camera}} Gene
Set Analysis (GSA) and visualize results in barplot (two classes) or
heatmap (>2 classes).
}
\examples{
# sample subset for reduced run-time
subset <- 25:75
cam <- subCamera(emat=crcTCGAsubset, class=crcTCGAsubset$CMS,
    rowCluster=FALSE, doVoom=TRUE, keepN=subset)
}
\references{
Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, et al. limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucl. Acids Res. 2015;gkv007.

Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics. 2010;26:139-40.

Law CW, Chen Y, Shi W, Smyth GK. voom: precision weights unlock linear model analysis tools for RNA-seq read counts. Genome Biology. 2014;15:R29.

Wu D, Smyth GK. Camera: a competitive gene set test accounting for inter-gene correlation. Nucleic Acids Res. 2012;gks461.
}
