% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subDEG.R
\name{subDEG}
\alias{subDEG}
\title{limma/voom differential gene expression analysis}
\usage{
subDEG(
  emat,
  class,
  batch = NULL,
  keepN = TRUE,
  doPairwise = FALSE,
  doVoom = FALSE,
  normMethod = "quantile",
  returnTreat = FALSE,
  lfc = log2(1.5),
  padj = 0.05,
  feat = NULL,
  doPlot = FALSE,
  sortBy = "P",
  ...
)
}
\arguments{
\item{emat}{numeric matrix with row features and sample columns. For
\code{\link[Biobase]{ExpressionSet}}
\code{\link[Biobase]{exprs}} matrix is used and featureData used for
annotating results.}

\item{class}{a factor vector specifying subtypes compared.
\code{length(class)==ncol(emat)}.}

\item{batch}{a factor vector specifying additional level in design matrix.}

\item{keepN}{a logical or numeric vector specifying which samples to keep
(defaults to all).}

\item{doPairwise}{a logical, whether pairwise comparison are to be
performed (as opposed to class k against K-k).}

\item{doVoom}{a logical, indicating whether emat is sequencing count data.}

\item{normMethod}{a character, only used if doVoom=TRUE and passed to}

\item{returnTreat}{logical, if true, parameters lfc and padj are passed to \code{\link[limma]{topTreat}}, else \code{\link[limma]{eBayes}} are used and results sorted as indicated by sortBy parameter.}

\item{lfc}{numeric, only applicable if returnTreat=TRUE, if FALSE, all genes are returned}

\item{padj}{numeric, only applicable if returnTreat=TRUE, if FALSE, all genes are returned}

\item{feat}{data frame with feature data}

\item{doPlot}{a logical, not implemented.}

\item{sortBy}{character passed to \code{\link[limma]{topTable}}
\code{sort.by} parameter.}

\item{...}{additional arguments passed to \code{\link[edgeR]{calcNormFactors}}
\code{\link[edgeR]{calcNormFactors}} if element in c("TMM","RLE",
"upperquartile","none") or \code{\link[limma]{voom}} ("scale", "quantile",
"cyclicloess").}
}
\value{
a data frame (for two classes) or list of data frames (in case of
more than two classes) with results.
}
\description{
Wrapper function to perform \code{\link[limma]{eBayes}} differential
gene expression analysis (DEG).
}
\examples{
deg <- subDEG(crcTCGAsubset, crcTCGAsubset$CMS, doVoom=TRUE, sortBy="P")
lapply(deg, head)
}
\references{
Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, et al. limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucl. Acids Res. 2015;gkv007. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4402510/}

Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics. 2010;26:139-40. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/pmid/19910308/}

Law CW, Chen Y, Shi W, Smyth GK. voom: precision weights unlock linear model analysis tools for RNA-seq read counts. Genome Biology. 2014;15:R29. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/pmid/24485249/}
}
