% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subHeatmap.R
\name{subHeatmap}
\alias{subHeatmap}
\title{ntp heatmap}
\usage{
subHeatmap(
  emat,
  res,
  templates,
  keepN = TRUE,
  labRow = NULL,
  classCol = getOption("subClassCol"),
  heatCol = NULL,
  ...
)
}
\arguments{
\item{emat}{a numeric matrix with row features and sample columns.}

\item{res}{a data frame, result from \code{ntp} function.}

\item{templates}{data frame with two columns; class (factor or numeric) and
probe (character).}

\item{keepN}{a logical or numeric vector specifying which samples to keep.}

\item{labRow}{a character specifying \code{templates} column used for labeling
features.}

\item{classCol}{a character vector specifying class colors.}

\item{heatCol}{a character vector specifying heatmap colors (must be an even
number).}

\item{...}{additional arguments passed to \code{\link[graphics]{image}}.}
}
\value{
a raster heatmap to visualize template feature expression.
}
\description{
Heatmap with marker genes and class labels.
}
\details{
\code{subHeatmap} provides a simple heatmap with samples and genes
ordered according to subtype. The column bar indicates prediction
\eqn{p}-values.
}
