% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subMA.R
\name{subMA}
\alias{subMA}
\title{MA plot}
\usage{
subMA(
  deg,
  geneID = "rownames",
  lfc = log2(2),
  padj = 0.01,
  ave = 1,
  topN = 10,
  cexText = 1,
  classCol = getOption("subClassCol"),
  ...
)
}
\arguments{
\item{deg}{output from \link{subDEG} which may be either a list or
data.frame.}

\item{geneID}{character, column names to use for labeling top hits.}

\item{lfc}{numeric absolute \eqn{log2(fold-change)} threshold.}

\item{padj}{numeric, adjusted \eqn{p}-value threshold.}

\item{ave}{numeric, average expression threshold.
\code{length(class)==ncol(emat)}.}

\item{topN}{an integer, number of genes to label in plot (selected by
largest absolute \eqn{fold-change}).}

\item{cexText}{numeric, scaling factor for labels for top hits.}

\item{classCol}{a character vector specifying class colors.}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}} or
\code{\link[graphics]{smoothScatter}} or if \pkg{KernSmooth} is available
(currently only main, xlab and ylab used).}
}
\value{
a MA-plot
}
\description{
Mean versus average (MA) plot of output from \link{subDEG}.
}
\examples{
deg <- subDEG(crcTCGAsubset, crcTCGAsubset$CMS, doVoom=TRUE, sortBy="B")
subMA(deg, geneID="symbol", lfc=3)
}
