% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subMDS.R
\name{subMDS}
\alias{subMDS}
\title{MDS plot with class labels}
\usage{
subMDS(
  emat,
  class = NULL,
  keepN = TRUE,
  nGenes = 1000,
  normMethod = NULL,
  dim = c(1, 2),
  labelSamp = FALSE,
  labelCenters = TRUE,
  classConfusion = NULL,
  legend = "bottomleft",
  classCol = getOption("subClassCol"),
  ...
)
}
\arguments{
\item{emat}{a numeric matrix with row features and sample columns.}

\item{class}{a factor vector specifying subtypes compared
\code{length(class)==ncol(emat)}.}

\item{keepN}{a logical or numeric vector specifying which samples to keep
(defaults to all).}

\item{nGenes}{an integer specifying number of features to used for MDS}

\item{normMethod}{a character, passed to \code{\link[edgeR]{calcNormFactors}}
if element in c("TMM","RLE", "upperquartile","none") or
\code{\link[limma]{voom}} ("scale", "quantile", "cyclicloess").}

\item{dim}{a numeric vector of length 2, specifying which principal
components to plot.}

\item{labelSamp}{a logical indicating whether points (FALSE) or
colnames(emat) are plotted.}

\item{labelCenters}{a logical indicating whether class centers should be
labeled.}

\item{classConfusion}{a factor vector with same length and levels as
\code{class} specifying alternative classifications.}

\item{legend}{a character specifying legend placement (e.g. "topleft",
"bottomright").}

\item{classCol}{a character vector of hexcolors with
\code{length(classCol)>=levels(class)}. If NULL, a pre-defined RColorBrewer
based palette is used.}

\item{...}{arguments to be passed to \code{\link[limma]{plotMDS}}.}
}
\value{
A labeled MDS plot
}
\description{
Multi-dimensional scaling (MDS) plot with class labels. Wrapper
for \code{\link[limma]{plotMDS}}.
}
\details{
\code{plotMDS} provides a MDS plot of emat with points colored
according to class labels. If \emph{classConfusion} is specified, samples
where \code{class != classConfusion} are highlighted.
}
\examples{
# sample subset for reduced run-time
subset <- 25:50
subMDS(emat=crcTCGAsubset, class=crcTCGAsubset$CMS,
        keepN=subset, normMethod="quantile")
}
