% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subPairs.R
\name{subPairs}
\alias{subPairs}
\title{scatterplots of sample-templates distances}
\usage{
subPairs(res, classCol = getOption("subClassCol"))
}
\arguments{
\item{res}{a data frame, result from \code{\link{ntp}} function.}

\item{classCol}{a character vector specifying class colors}
}
\value{
a panel of plots of pair-wise template-distances.
}
\description{
Pairwise sample-templates distance plots with samples colored
according to predicted class.
}
\details{
Upper-panels show the Pearson correlations
between sample-templates distances. Diagonal panels show class and histogram
of distances. The vertical, dashed lines are the distance between two
uncorrelated vectors.
}
\note{
smaller distances indicate more similar samples.
}
\examples{
 res <- CMScaller(crcTCGAsubset, nPerm=1000, RNAseq=TRUE)
 subPairs(res)
}
\seealso{
\code{\link{ntp}}, \code{\link{corCosine}}, \code{\link{pairs}}
}
