% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subPCA.R
\name{subPCA}
\alias{subPCA}
\title{PCA plot with class labels.}
\usage{
subPCA(
  emat,
  class = NULL,
  keepN = TRUE,
  nGenes = 1000,
  normMethod = NULL,
  dim = c(1, 2),
  labelSamp = FALSE,
  legend = "topright",
  classConfusion = NULL,
  classCol = getOption("subClassCol"),
  ...
)
}
\arguments{
\item{emat}{numeric matrix with row features and sample columns.}

\item{class}{a factor vector specifying sample classes
\code{length(class)==ncol(emat)}.}

\item{keepN}{a logical or numeric vector specifying which samples to keep.}

\item{nGenes}{an integer specifying number of features to used for PCA.}

\item{normMethod}{a character, passed to \code{\link[edgeR]{calcNormFactors}}
if element in c("TMM","RLE", "upperquartile","none") or
\code{\link[limma]{voom}} ("scale", "quantile", "cyclicloess").#'}

\item{dim}{a numeric vector of length 2, specifying which principal
components to plot.}

\item{labelSamp}{a logical indicating whether points or
\code{colnames(emat)} are plotted.}

\item{legend}{a character specifying legend placement (e.g. "topleft",
"bottomright").}

\item{classConfusion}{a factor vector with same length and levels as
\code{class} specifying alternative classifications.}

\item{classCol}{a character vector of hexcolors with
\code{length(classCol)>=levels(class)}.}

\item{...}{arguments to be passed to \code{\link[graphics]{plot}}.}
}
\value{
a labeled PCA plot and list of class \code{prcomp} passed
from \code{\link[stats]{prcomp}}
}
\description{
Principal component analysis (PCA) plot with class labels.
}
\details{
\code{plotPCA} provides a PCA with points colored
according to class labels. If \code{classConfusion} is specified, samples
where \code{class != classConfusion} are highlighted.
}
\seealso{
\code{\link[stats]{prcomp}}, \code{\link{subMDS}}
}
