% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSetNA.R
\name{subSetNA}
\alias{subSetNA}
\title{sets NTP results to NA based on p-value cutoff}
\usage{
subSetNA(res, pValue = 1, FDR = 1, verbose = TRUE)
}
\arguments{
\item{res}{a data frame, result from \code{\link{ntp}} function.}

\item{pValue}{a numeric, predictions with higher p-values are set to NA.}

\item{FDR}{a numeric, predictions with higher FDR are set to NA.}

\item{verbose}{logical, whether console messages are to be displayed.}
}
\value{
A data.frame where res$prediction's are set to NA based on
\eqn{p}-value or FDR threshold.
}
\description{
sets \code{ntp} predictions to NA based on \eqn{p}-value or FDR
thresholds
}
\details{
Replaces low-confidence predictions with NA.
}
\examples{
emat <- ematAdjust(crcTCGAsubset, normMethod = "quantile")
res <- ntp(emat, templates.CMS, doPlot=TRUE, nPerm=100)
res <- subSetNA(res, pValue=.1)
res <- subSetNA(res, FDR=.1)
table(res$prediction, useNA="always")
}
