% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subVolcano.R
\name{subVolcano}
\alias{subVolcano}
\title{volcano plot}
\usage{
subVolcano(
  deg,
  geneID = "rownames",
  lfc = log2(2),
  padj = 0.05,
  ave = 0,
  topN = 10,
  cexText = 1,
  classCol = getOption("subClassCol"),
  ...
)
}
\arguments{
\item{deg}{output from \code{\link{subDEG}} which may be either a list or
a data frame.}

\item{geneID}{\code{deg} column name to use for labeling top hits.}

\item{lfc}{a numeric absolute \eqn{log2(fold-change)} threshold.}

\item{padj}{a numeric, adjusted \eqn{p}-value threshold.
\code{length(class)==ncol(emat)}.}

\item{ave}{a numeric, average expression threshold.}

\item{topN}{an integer, number of genes to label in plot (selected by
largest absolute \eqn{fold-change}).}

\item{cexText}{a numeric, scaling factor for labels for top hits.}

\item{classCol}{a character vector specifying class colors.}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}} or
\code{\link[graphics]{smoothScatter}} or if \pkg{KernSmooth} is available
(currently only main, xlab and ylab used).}
}
\value{
one or more volcano plots where horizontal dashed line is the maximum
\eqn{p}-value below \code{padj} and vertical lines the \code{lfc}-threshold.
}
\description{
Volcano plot of output from \code{\link{subDEG}}.
}
