% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{templates.CRIS}
\alias{templates.CRIS}
\title{CRC intrinsic subtypes (CRIS) templates}
\format{
An object of class \code{data.frame} with 565 rows and 3 columns.
}
\usage{
templates.CRIS
}
\description{
CRC intrinsic subtypes (CRIS) templates
}
\details{
Colorectal cancer CRC intrinsic subtypes (CRIS) prediction
templates for \code{\link[CMScaller]{ntp}} retrieved from Supplementary Table
8 (Isella 2017). Suspicious (Mar-03) or legacy gene symbols were updated.
}
\examples{
dim(templates.CRIS)
}
\references{
Isella C, Brundu F, Bellomo SE, Galimi F, Zanella E, Porporato R, et al. Selective analysis of cancer-cell intrinsic transcriptional traits defines novel clinically relevant subtypes of colorectal cancer. Nat Commun [Internet]. 2017 [cited 2017 Sep 11];8:ncomms15107. Available from: \url{https://www.nature.com/articles/ncomms15107}
}
\keyword{datasets}
