% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{templates.MSI}
\alias{templates.MSI}
\title{micro-satellite instability templates}
\format{
An object of class \code{data.frame} with 100 rows and 3 columns.
}
\usage{
templates.MSI
}
\description{
micro-satellite instability templates
}
\details{
micro-satellite instability (MSI) prediction templates for
\code{\link[CMScaller]{ntp}}. Marker genes were identified using TCGA
RNA-sequencing data. \code{templates$probe} refers to Entrez ids.
}
\examples{
head(templates.MSI)
table(templates.MSI$class)
}
\keyword{datasets}
