% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno.orgHs}
\alias{anno.orgHs}
\title{annotation for translating between gene identifiers}
\format{
An object of class \code{data.frame} with 58219 rows and 7 columns.
}
\usage{
anno.orgHs
}
\description{
Table for translating between Ensembl, NCBI Entrez and HGNC symbol
human gene identifiers. Gene biotype, exonic gene length and proportion
GC-bases are also included. For more exhaustive annotation including gene
names, GO terms and so on, please refer to R Bioconductor packages
\code{AnnotationDbi} and \code{org.Hs.eg.db}.
}
\details{
Data is from GENCODE v26/GRCh38.p10. with NCBI Entrez identifiers matched using HGNC symbol against \code{\link[org.Hs.eg.db]{org.Hs.eg.db}} package.
}
\examples{
head(anno.orgHs)
dim(anno.orgHs)
colSums(apply(anno.orgHs, 2, is.na))
colSums(!apply(anno.orgHs[,1:4], 2, duplicated))
# some extremes
anno.orgHs[which.max(anno.orgHs$length),]
anno.orgHs[which.min(anno.orgHs$length),]
anno.orgHs[which.max(anno.orgHs$GC),]
anno.orgHs[which.min(anno.orgHs$GC),]
attributes(anno.orgHs)$genome
}
\references{
Carlson M. org.Hs.eg.db: Genome wide annotation for Human. R package version 3.7.0.

Harrow J, Frankish A, Gonzalez JM, Tapanari E, Diekhans M, Kokocinski F, et al. GENCODE: The reference human genome annotation for The ENCODE Project. Genome Res. 2012;22:1760-74.
}
\seealso{
\code{\link{fromTo}}, \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
}
\keyword{datasets}
