% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSgsa.R
\name{CMSgsa}
\alias{CMSgsa}
\title{CMS Gene Set Analysis}
\usage{
CMSgsa(emat, class, RNAseq = FALSE, ...)
}
\arguments{
\item{emat}{a numeric expression matrix with sample columns, gene rows and
Entrez rownames. Microarray data should be normalized and log2-transformed.
For RNA-seq data, raw counts or RSEM values could be used directly by setting
\code{RNAseq=TRUE}.}

\item{class}{a factor vector specifying sample classes.}

\item{RNAseq}{a logical, set to TRUE if emat is untransformed, non-normalized
sequencing counts or RSEM values.}

\item{...}{additional arguments passed to \code{\link{subCamera}}.}
}
\value{
a heatmap and \code{\link[limma]{camera}} output (list,
invisible). In heatmap, red and blue indicates relative up- and
down-regulation respectively. Color saturation reflects significance.
Nominal `camera` \eqn{p}-values are used as input for visualization.
}
\description{
\code{\link[limma]{camera}} gene set analysis (GSA) against 15
pre-selected CMS-informative gene sets
(\code{\link{geneSets.CMS}}).
}
\details{
See \code{\link{subCamera}} for output details.
}
\examples{
cam <- CMSgsa(emat=crcTCGAsubset, class=crcTCGAsubset$CMS, RNAseq=TRUE)
lapply(cam, head)
}
\references{
Guinney J, Dienstmann R, Wang X, de Reyniès A, Schlicker A, Soneson C, et al. The consensus molecular subtypes of colorectal cancer. Nat Med. 2015;21:1350–6. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4636487/}

Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, et al. limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucl. Acids Res. 2015;gkv007. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4402510/}

Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics. 2010;26:139-40. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/pmid/19910308/}

Law CW, Chen Y, Shi W, Smyth GK. voom: precision weights unlock linear model analysis tools for RNA-seq read counts. Genome Biology. 2014;15:R29. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/pmid/24485249/}
}
\seealso{
\code{\link[limma]{camera}}, \code{\link{subCamera}}, \code{\link{geneSets.CMS}}
}
