% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{templates.CMS}
\alias{templates.CMS}
\title{development consensus molecular subtype (CMS) templates}
\format{
An object of class \code{data.frame} with 529 rows and 3 columns.
}
\usage{
templates.CMS
}
\description{
development consensus molecular subtype (CMS) templates
}
\details{
Colorectal cancer Consensus Molecular Subtypes (CMS) prediction
templates for \code{\link[CMScaller]{ntp}}. Marker genes were
identified using TCGA RNA-sequencing data. \code{templates$probe} refers to
Entrez ids. Compared to \code{\link{templates.CMS}} this was prepared using
additional filters steps to further clean genes also expressed by
non-carcinoma cell types. Predictions are usually highly concordant with
\code{\link{templates.CMS}}. Under development.
}
\examples{
length(intersect(templates.CMS$symbol, templates.CMS$symbol))
length(setdiff(templates.CMS$symbol, templates.CMS$symbol))
}
\references{
Guinney J, Dienstmann R, Wang X, de Reyniès A, Schlicker A, Soneson C, et al. The consensus molecular subtypes of colorectal cancer. Nat Med. 2015;21:1350–6. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4636487/}

Hoshida Y. Nearest Template Prediction: A Single-Sample-Based Flexible Class Prediction with Confidence Assessment. PLoS ONE. 2010;5:e15543. Available from: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/pmid/21124904/}
}
\keyword{datasets}
